requirements.txt
reportlab>=3.6.8
PyPDF2>=3.0.0
Pillow>=9.0.0
pdf2image>=1.16.0
pyinstaller>=5.0.0 